#157: [Reafactor] componentごとにファイル分け
