+++ frontend-nodejs/src/app/(auth)/login/components/RequiredTag.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export const RequiredTag = () => {
+  return (
+    <span className="ml-2 px-2 py-0.5 bg-amber-600 text-white text-xs font-bold rounded-full animate-pulse">
+      必須
+    </span>
+  );
+};
