+++ frontend-nodejs/src/app/(auth)/login/components/WoodenSign.tsx
@@ -1,0 +1,58 @@
+import React, { ReactNode } from "react";
+
+export const WoodenSign = ({
+  children,
+  width = "w-64",
+  height = "h-12",
+  rotation = "",
+}: {
+  children: ReactNode;
+  width?: string;
+  height?: string;
+  rotation?: string;
+}) => {
+  // 釘コンポーネント
+  const Nail = ({
+    position,
+  }: {
+    position: "topLeft" | "topRight" | "bottomLeft" | "bottomRight";
+  }) => {
+    const positionClasses = {
+      topLeft: "top-1 left-1",
+      topRight: "top-1 right-1",
+      bottomLeft: "bottom-1 left-1",
+      bottomRight: "bottom-1 right-1",
+    };
+
+    return (
+      <div
+        className={`absolute ${positionClasses[position]} w-1.5 h-1.5 rounded-full bg-gray-500 border border-gray-500 shadow-inner`}
+        style={{ boxShadow: "inset 0 0 2px rgba(255,255,255,0.5)" }}
+      />
+    );
+  };
+
+  return (
+    <div className={`relative ${width}`}>
+      {/* 影の要素 - 看板とサイズを合わせる */}
+      <div
+        className={`absolute w-full ${height} top-[5px] left-[6px] rounded`}
+        style={{
+          backgroundColor: "rgba(0,0,0,0.7)",
+          filter: "blur(2px)",
+          zIndex: 5,
+        }}
+      />
+
+      <div
+        className={`relative ${width} ${height} bg-orange-300 flex items-center justify-center px-4 transform ${rotation} border-2 border-yellow-900 rounded z-10`}
+      >
+        <Nail position="topLeft" />
+        <Nail position="topRight" />
+        <Nail position="bottomLeft" />
+        <Nail position="bottomRight" />
+        {children}
+      </div>
+    </div>
+  );
+};
